rename ProofsCard to Card
