+++ src/components/proofs/Card.tsx
